+++ core/synthesis/structure_synthesis/__init__.py
@@ -18,7 +18,7 @@ from typing import (
     Dict,
     Optional,
 )
-from time import time
+from time import perf_counter
 from pyslvs import (
     Graph,
     link_assortment,
@@ -332,7 +332,7 @@ class StructureSynthesis(QWidget, Ui_Form):
 
     def __structural_combine(self, jobs: Sequence[QTreeWidgetItem]) -> None:
         """Structural combine by iterator."""
-        t0 = time()
+        t0 = perf_counter()
         dlg = SynthesisProgressDialog(
             "Structural Synthesis",
             f"Number of cases: {len(jobs)}",
@@ -360,7 +360,7 @@ class StructureSynthesis(QWidget, Ui_Form):
                     except ValueError:
                         pass
                 root.setText(1, f"{count}")
-            self.__set_time_count(time() - t0, len(self.answer))
+            self.__set_time_count(perf_counter() - t0, len(self.answer))
             self.__reload_atlas()
 
         work = GraphEnumerateThread(jobs, self.graph_degenerate.currentIndex(), dlg)
@@ -391,7 +391,7 @@ class StructureSynthesis(QWidget, Ui_Form):
             self
         )
         dlg.show()
-        t0 = time()
+        t0 = perf_counter()
         for i, G in enumerate(self.answer):
             QCoreApplication.processEvents()
             if dlg.wasCanceled():
@@ -400,7 +400,7 @@ class StructureSynthesis(QWidget, Ui_Form):
                 dlg.setValue(i + 1)
             else:
                 break
-        self.__set_paint_time(time() - t0)
+        self.__set_paint_time(perf_counter() - t0)
         dlg.setValue(dlg.maximum())
         dlg.deleteLater()
         scroll_bar.setSliderPosition(scroll_pos)